Allow paged resources to be interated through without making the Namely API timeout.
This allows resources that sometimes contain a TON of data (IE: Profiles) to be paged through. Currently this client uses the limit=all parameter when making requests. For companies with mass amounts of profiles, this breaks pretty fast. This makes a change for profiles that profiles are paged through and iterated out sequentially.